requirements.txt
pygments
numpy
scipy
matplotlib
pyqt6
pyudev
PyOpenGL PyOpenGL_accelerate